requirements.txt
requests
Pillow

